backend/requirements.txt
Django==2.2.6
django-choices==1.7.0
django-extensions==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-gis==0.14
GDAL==2.3.2
psycopg2==2.8.4
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
